Folders/themes draw in grids + rebooting indicator
Folders and themes now draw in grids instead of the ugly button lists.
They will show default icons with the first letter of the folder/theme
on top, or if folder.smdh or theme.smdh is placed inside the
folder/theme directory then the image inside this smdh will be used as
the icon for the grid. This change also removes the 8 folder/theme
limit, as the folder/theme grids inherit the paging ability of the
other grids so can show as many themes/folders as are on the device.

In addition, the “Reboot” icon now changes its text to “Rebooting…”
when it has been activated so the user will know the device is in the
process of rebooting.

Finally, I have created a spinning progress wheel. This isn’t used
anywhere yet but is included in this commit.